Update code for chisel 3.3 Refactored most things. Down to just a couple of fails that appear to be a bit more complicated.
