+++ src/main/scala/jtag/Utils.scala
@@ -12,8 +12,7 @@ class Tristate extends Bundle {
   val driven = Bool()  // active high, pin is hi-Z when driven is low
 }
 
-class NegativeEdgeLatch[T <: Data](clock: Clock, dataType: T)
-    extends Module(override_clock=Some(clock)) {
+class NegativeEdgeLatch[T <: Data](dataType: T) extends Module {
   class IoClass extends Bundle {
     val next = Input(dataType)
     val enable = Input(Bool())
@@ -31,9 +30,12 @@ class NegativeEdgeLatch[T <: Data](clock: Clock, dataType: T)
 /** Generates a register that updates on the falling edge of the input clock signal.
   */
 object NegativeEdgeLatch {
-  def apply[T <: Data](clock: Clock, next: T, enable: Bool=true.B): T = {
+  def apply[T <: Data](modClock: Clock, next: T, enable: Bool=true.B): T = {
     // TODO better init passing once in-module multiclock support improves
-    val latch_module = Module(new NegativeEdgeLatch((!clock.asUInt).asClock, next.cloneType))
+
+    val latch_module = withClock((!(modClock.asUInt)).asClock) {
+      Module(new NegativeEdgeLatch(next.cloneType))
+    }
     latch_module.io.next := next
     latch_module.io.enable := enable
     latch_module.io.output
@@ -43,8 +45,7 @@ object NegativeEdgeLatch {
 /** A module that counts transitions on the input clock line, used as a basic sanity check and
   * debug indicator clock-crossing designs.
   */
-class ClockedCounter(modClock: Clock, counts: BigInt, init: Option[BigInt])
-    extends Module(override_clock=Some(modClock)) {
+class ClockedCounter(counts: BigInt, init: Option[BigInt]) extends Module {
   require(counts > 0, "really?")
 
   val width = log2Ceil(counts)
@@ -71,11 +72,15 @@ class ClockedCounter(modClock: Clock, counts: BigInt, init: Option[BigInt])
   */
 object ClockedCounter {
   def apply (data: Bool, counts: BigInt, init: BigInt): UInt = {
-    val counter = Module(new ClockedCounter(data.asClock, counts, Some(init)))
+    val counter = withClock(data.asClock) {
+      Module(new ClockedCounter(counts, Some(init)))
+    }
     counter.io.count
   }
   def apply (data: Bool, counts: BigInt): UInt = {
-    val counter = Module(new ClockedCounter(data.asClock, counts, None))
+    val counter = withClock(data.asClock) {
+      Module(new ClockedCounter(counts, None))
+    }
     counter.io.count
   }
 }
